### 2019-03-19 #290 @cstubben
Submission: tidypmc
Submitting Author: Chris Stubben (@cstubben)
Repository:  https://github.com/cstubben/tidypmc
Version submitted:   1.1
Version accepted: TBD

---

```
Package: tidypmc
Title: Parse full text XML documents from PMC
Version: 1.1
Author: Chris Stubben
Maintainer: Chris Stubben <chris.stubben@hci.utah.edu>
Depends: xml2
    europepmc,
    tokenizers,
    readr
    tidytext,
    knitr
Description: This package parses section paragraphs, captions, tables, references and metadata from XML documents in the Open Access subset of Pubmed Central. Additional functions are available to search text and expand ranges of the references cited, locus tags and gene operons.
```

`tidypmc` parses paragraphs, captions, tables and references in XML documents from Pubmed Central into tidy formats for searching and further analysis with `tidytext`

Researchers that want to analyze full text documents in PMC.
`pubchunks`.   `pub_chunks(x, "body")` returns a list with an s3 class and this output can be converted to a data.frame, but the document structure is lost and text often contains nested tables, mathML and other text blobs.  `pmc_text` removes tables and other tags [incorrectly] nested in //sec/p tags and tries to preserve the document structure by saving the full path to subsection titles, paragraph numbers and sentence in a tibble.
