+++ R/pmc_caption.R
@@ -30,7 +30,7 @@ pmc_caption <- function(doc){
       # in bold tags or have very long titles that should be split.
       ## use node() or * to avoid pasting /title and /p sentences without a space ()
       f2 <- sapply(z, function(x) paste(xml_text(xml_find_all(x, "./caption/*")), collapse=" "))
-      names(f2) <- f1
+      names(f2) <- gsub("\\.$", "", f1)
       x1 <- sapply(f2, tokenizers::tokenize_sentences)
       figs <- dplyr::bind_rows(
          lapply(x1, function(z) tibble::tibble(sentence = 1:length(z), text=z)), .id="label")
@@ -46,7 +46,7 @@ pmc_caption <- function(doc){
       f1 <- sapply(z, function(x) xml_text(xml_find_first(x, "./label"), trim=TRUE))
       # some with long subcaptions
       f2 <- sapply(z, function(x) paste( xml_text(xml_find_all(x, "./caption/*")), collapse=" "))
-      names(f2) <- f1
+      names(f2) <- gsub("\\.$", "", f1)
       x1 <- sapply(f2, tokenizers::tokenize_sentences)
       tbls <- dplyr::bind_rows(
          lapply(x1, function(z) tibble::tibble(sentence = 1:length(z), text=z)), .id="label")
@@ -60,16 +60,20 @@ pmc_caption <- function(doc){
       message("Found ", n, ifelse(n>1, " supplements", " supplement"))
       ## BMC has labels in caption title (Additional file 1) and start of paragraph (Figure S1)
       f1 <- sapply(z, function(x) xml_text(xml_find_first(x, "./label"), trim=TRUE))
-      ## use ./caption to avoid media/caption tags...
+      ## use ./caption to avoid extra media/caption tags???
       f2 <- sapply(z, function(x) paste( xml_text(xml_find_all(x, "./caption/*")), collapse=" "))
-      # mBio with /p tags only
-      if(all(f2 == "")) f2 <- sapply(z, function(x) xml_text(xml_find_all(x, "./p")))
+      # mBio with /p tags only, others with media/caption only
+      if(all(is.na(f1)) & all(f2=="")){
+         ## ANY label and ANY paragrah
+         f1 <- sapply(z, function(x) xml_text(xml_find_first(x, ".//label"), trim=TRUE))
+         f2 <- sapply(z, function(x) xml_text(xml_find_all(x, ".//p")))
+      }
       #  text in media/ tag
       if(is.na(f1[1]) & length(f2[[1]]) == 0){
          message(" No supplement /caption or /p tag to parse")
          sups <- NULL
       }else{
-         ## remove period to avoid splitting (DOC), (XLSX), etc into new sentence.
+         ## remove period to avoid splitting (DOC), (XLSX) into new sentences - misses (XLSX 32 kb) 
          f2 <- gsub("\\.( \\([A-Z]+\\))", "\\1", f2)
          x1 <- sapply(f2, tokenizers::tokenize_sentences, USE.NAMES=FALSE)
          if(all(is.na(f1))){
@@ -79,11 +83,12 @@ pmc_caption <- function(doc){
                 f1 <- y
                 x1 <- sapply(x1, function(x) x[-1])
             }else{
-               message(" Missing supplement label tag, using File S1 to ", length(z))
+               if(length(z) == 1) message(" Missing supplement label tag, using File S1")
+               else message(" Missing supplement label tag, using File S1 to S", length(z))
                f1 <- paste0("File S", 1:length(z))
             }
          }
-         names(x1) <- f1
+         names(x1) <- gsub("\\.$", "", f1)
          sups <- dplyr::bind_rows(
             lapply(x1, function(z) tibble::tibble(sentence = 1:length(z), text=z)), .id="label")
       }
