### 2019-03-20 #290 @melvidoni
Hello @cstubben,  thank you for your submission.

On the Scope section of the opening post, could you explain possible differences and similitudes between your package and  [pubchunks](https://github.com/ropensci/pubchunks/) which is already onboarded to rOpenSci?
