### 2019-03-20 #290 @cstubben
Thanks for the suggestion.  pubchunks has now been added.
