+++ man/separate_refs.Rd
@@ -15,7 +15,7 @@ separate_refs(txt, column = "text")
 a tibble
 }
 \description{
-Separates references cited in brackets or parenthesis into multiple rows and splits
+Separates references cited in brackets or parentheses into multiple rows and splits
 the comma-delimited numeric strings and expands ranges like 7-9 into new rows
 }
 \examples{
