### 2019-03-22 #290 @noamross
### Editor checks:

---

#### Editor comments

Hello @cstubben, thanks for your submission! This package is a good fit under our Aims and Scope.  However, we require that packages have test suite with good coverage of package functionality and continuous integration before we pass them on to review.  You can find more details here: <https://ropensci.github.io/dev_guide/building.html#testing>

In addition, here's the output of `goodpractice::gp()` run on your package.  You should address these notes as well before we move to review.

Please let me know if you have questions. Once you have addressed these issues we can move to assigning reviewers.

```
── GP tidypmc ──────────────────────────────────

It is good practice to

    R/collapse_rows.R:20:NA
    R/collapse_rows.R:21:NA
    R/collapse_rows.R:22:NA
    R/collapse_rows.R:23:NA
    R/collapse_rows.R:24:NA
    ... and 398 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name
    clashes, and poor interaction with other packages. Use "Imports"
    instead.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/separate_refs.R:20:11

    R/pmc_caption.R:28:1
    R/pmc_caption.R:29:1
    R/pmc_caption.R:31:1
    R/pmc_caption.R:32:1
    R/pmc_caption.R:35:1
    ... and 81 more lines

    vapply() instead.

    R/collapse_rows.R:29:13
    R/pmc_caption.R:28:13
    R/pmc_caption.R:32:13
    R/pmc_caption.R:35:16
    R/pmc_caption.R:36:16
    ... and 31 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/collapse_rows.R:40:32
    R/pmc_caption.R:48:62
    R/pmc_caption.R:67:59
    R/pmc_caption.R:111:39
    R/pmc_caption.R:116:62    ... and 12 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
```
