+++ R/pmc_table.R
@@ -23,20 +23,19 @@
 #' @export
 
 pmc_table  <- function(doc){
-   z  <- xml_find_all(doc, "//table-wrap")
+    twn  <- length(xml_find_all(doc, "//table-wrap"))
+   ## Avoid table-wrap without table node .. link to image only
+   z  <- xml_find_all(doc, "//table-wrap/table/..")
    if (length(z) == 0) {
       message("No tables found")
+      if(twn > 0) message("Table-wrap with link to image?")
       tbls <- NULL
    }else{
-      tbl_nodes <- xml_find_all(z, "./table")
-      ## some table tags are missing
-      ## SEE https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2211553/table/ppat-0040009-t001/
-      if(length(tbl_nodes)==0){
-         message("Found ", length(z), " /table-wrap tags")
-         message(" No /table tags found - possible link to image?")
-         tbls <- NULL
-      }else{
+        tbl_nodes <- xml_find_all(z, "./table")
         message("Found ", length(z), " tables")
+        if(twn > length(z)) message(twn-length(n), " /table-wrap with link to image?")
+        ## START table function
+        #  t1 <- xml_find_all(doc, "//table")[1]
         tbls <- lapply(tbl_nodes, function(t1){
            #PARSE HEADER
            x <- xml_find_all(t1, ".//thead/tr")
@@ -53,10 +52,13 @@ pmc_table  <- function(doc){
                 thead <- rep(thead, colspan)
               }
            # mutliline header - collapse into single row
-           # SEE  tables 1 and 2 in https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3109299
+           # SEE  tables 1 and 2 in PMC3109299
            }else{
                nr <- length(x)
-               c2 <- data.frame()
+               nc <- max(sapply(x, function(y) sum(as.numeric(
+                   xml_attr( xml_find_all(y, ".//td|.//th"), "colspan", default="1")))))
+
+               c2 <- data.frame(matrix(NA, nrow = nr , ncol =  nc ))
                for (i in 1:nr){
                   rowspan <- as.numeric( xml_attr( xml_find_all(x[[i]], ".//td|.//th"), "rowspan", default="1"))
                   colspan <- as.numeric( xml_attr( xml_find_all(x[[i]], ".//td|.//th"), "colspan", default="1"))
@@ -64,11 +66,6 @@ pmc_table  <- function(doc){
                   if( any(colspan>1) ){
                      thead   <- rep(thead, colspan)
                      rowspan <- rep(rowspan, colspan)
-                  }
-                  ## create empty data.frame
-                  if(i == 1){
-                     nc <- length(thead)
-                     c2 <- data.frame(matrix(NA, nrow = nr , ncol =  nc ))
                   }
                   # fill values into empty cells
                   n <- which(is.na(c2[i,]))
@@ -101,13 +98,17 @@ pmc_table  <- function(doc){
             x <- xml_find_all(t1, ".//tbody/tr")
             # number of rows
             nr <- length(x)
+            nc <- max(sapply(x, function(y) sum(as.numeric(
+                xml_attr( xml_find_all(y, ".//td|.//th"), "colspan", default="1")))))
+            c2 <- data.frame(matrix(NA, nrow = nr , ncol =  nc ))
+
             for (i in 1:nr){
                ## some table use //th  see table1 PMC3031304
                rowspan <- as.numeric( xml_attr( xml_find_all(x[[i]], ".//td|.//th"), "rowspan", default="1"))
[…]
