+++ R/pmc_reference.R
@@ -7,7 +7,7 @@
 #'
 #' @author Chris Stubben
 #'
-#' @note Mixed citations without any child tags are added to the title column.
+#' @note Mixed citations without any child tags are added to the author column.
 #'
 #' @examples
 #' # doc <- europepmc::epmc_ftxt("PMC2231364")
@@ -26,12 +26,12 @@ pmc_reference <- function (doc){
    x <- table(x[!x %in% c("label", "note")])
    message( "Found ", paste(x, names(x), collapse = " and "), " tags")
    ## xml_find_first returns NA for missing values
-   pmid <-  sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='pmid']"), trim=TRUE))
-   doi <-   sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='doi']"), trim=TRUE))
-   a1 <-    lapply(z, function(x) xml_text(xml_find_all(x, ".//surname"), trim=TRUE))
-   a2 <-    lapply(z, function(x) xml_text(xml_find_all(x, ".//given-names"), trim=TRUE))
+   pmid <- sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='pmid']"), trim=TRUE))
+   doi <-  sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='doi']"), trim=TRUE))
+   a1 <-   lapply(z, function(x) xml_text(xml_find_all(x, ".//surname"), trim=TRUE))
+   a2 <-   lapply(z, function(x) xml_text(xml_find_all(x, ".//given-names"), trim=TRUE))
    ## if all references have same number of authors, use SIMPLIFY=FALSE, see PMC6369050
-   authors<-sapply( mapply(paste, a1,a2, SIMPLIFY=FALSE), paste, collapse=", ")
+   authors <- sapply( mapply(paste, a1,a2, SIMPLIFY=FALSE), paste, collapse=", ")
    authors[authors == ""]<-NA
    # same authors published twice in same year, 2012a 2012b
    # year <-  sapply(z, function(x) xml_integer(xml_find_first(x, ".//year")))
