+++ R/pmc_metadata.R
@@ -18,7 +18,7 @@
 pmc_metadata <-function(doc ){
    z <- vector("list")
    ## //front has journal-meta and article-meta
-   # cat( as.character(xml_find_all(doc, "//journal-meta")))
+   # cat( as.character(xml_find_all(doc, "//front//journal-meta")))
    pmcid <- xml_text(xml_find_first(doc, "//front//article-id[@pub-id-type='pmcid']"))
    if(!is.na(pmcid))   z[["PMCID"]] <- paste0("PMC", pmcid)
    z[["Title"]] <- xml_text(xml_find_first(doc, "//front//article-title"))
@@ -30,38 +30,38 @@ pmc_metadata <-function(doc ){
    authors <-  paste(authors, collapse=", ")
    z[["Authors"]] <- authors
    ## Year published,  use collection else ppub year?
-   year <- xml_text( xml_find_first(doc, "//pub-date[@pub-type='collection']/year") )
-   if(is.na(year)) year <- xml_text( xml_find_first(doc, "//pub-date[@pub-type='ppub']/year") )
+   year <- xml_text( xml_find_first(doc, "//front//pub-date[@pub-type='collection']/year") )
+   if(is.na(year)) year <- xml_text( xml_find_first(doc, "//front//pub-date[@pub-type='ppub']/year") )
    if(!is.na(year))  z[["Year"]] <- year
 
    # Journal meta
-   journal <- xml_text( xml_find_first(doc,  "//journal-meta//journal-title"))
+   journal <- xml_text( xml_find_first(doc,  "//front//journal-meta//journal-title"))
    if(!is.na(journal))  z[["Journal"]] <- journal
    ## volume and issue in article metadata
-   volume <- xml_text( xml_find_first(doc, "//article-meta/volume"))
+   volume <- xml_text( xml_find_first(doc, "//front//article-meta/volume"))
    if(!is.na(volume)) z[["Volume"]] <- volume
-   issue <- xml_text( xml_find_first(doc, "//article-meta/issue"))
+   issue <- xml_text( xml_find_first(doc, "//front//article-meta/issue"))
    if(!is.na(issue)) z[["Issue"]] <- issue
    #PAGES
-   p1 <- xml_text( xml_find_first(doc, "//article-meta/fpage"))
+   p1 <- xml_text( xml_find_first(doc, "//front//article-meta/fpage"))
    if(!is.na(p1)){
-      p2 <- xml_text( xml_find_first(doc, "//article-meta/lpage"))
+      p2 <- xml_text( xml_find_first(doc, "//front//article-meta/lpage"))
       if(p1 != p2) p1 <- paste(p1, p2, sep="-")
    }else{
-      p1 <- xml_text( xml_find_first(doc, "//article-meta/elocation-id"))
+      p1 <- xml_text( xml_find_first(doc, "//front//article-meta/elocation-id"))
 
    }
    z[["Pages"]]  <- p1
    # More PUB Dates  - tags always sorted day, month, year?
-   epub <- xml_text( xml_find_all(doc, "//pub-date[@pub-type='epub']/*") )
+   epub <- xml_text( xml_find_all(doc, "//front//pub-date[@pub-type='epub']/*") )
    if(length(epub) > 0) z[["Published online"]] <- paste(rev(epub), collapse="-")
-   rec <- xml_text( xml_find_all(doc, "//history/date[@date-type='received']/*") )
+   rec <- xml_text( xml_find_all(doc, "//front//history/date[@date-type='received']/*") )
    if(length(rec) > 0) z[["Date received"]] <- paste(rev(rec), collapse="-")
    ## DOI
[…]
