### 2019-04-01 #290 @cstubben
Yes, I'm still trying to figure out the best way to do the unit tests.  I'd like to loop through 10K articles in one of the XML files at [Europe PMC FTP](https://europepmc.org/ftp/oa) rather than just test a single article if that's possible.
