### 2019-04-02 #290 @noamross
Good to hear.  If you have questions about topics like testing, the rOpenSci [forum](https://discuss.ropensci.org/) is a good place to ask.
