+++ R/pmc_reference.R
@@ -17,47 +17,67 @@
 #'
 #' @export
 
-pmc_reference <- function (doc){
+pmc_reference <- function(doc){
+   if(class(doc)[1] != "xml_document"){
+        stop("doc should be an xml_document from PMC")
+   }
    z <- xml_find_all(doc, "//ref")
    # cat(as.character(z[[1]]))
    if(length(z) > 0){
-   n <- sapply(z, function(x) xml_name(xml_find_all(x, "./*")))
+   n <- vapply(z, function(x) xml_name(xml_find_all(x, "./*")), character(1))
    x <- as.vector(unlist(n))
    x <- table(x[!x %in% c("label", "note")])
    message( "Found ", paste(x, names(x), collapse = " and "), " tags")
    ## xml_find_first returns NA for missing values
-   pmid <- sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='pmid']"), trim=TRUE))
-   doi <-  sapply(z, function(x) xml_text(xml_find_first(x, ".//pub-id[@pub-id-type='doi']"), trim=TRUE))
-   a1 <-   lapply(z, function(x) xml_text(xml_find_all(x, ".//surname"), trim=TRUE))
-   a2 <-   lapply(z, function(x) xml_text(xml_find_all(x, ".//given-names"), trim=TRUE))
-   ## if all references have same number of authors, use SIMPLIFY=FALSE, see PMC6369050
-   authors <- sapply( mapply(paste, a1,a2, SIMPLIFY=FALSE), paste, collapse=", ")
+   pmid <- vapply(z, function(x) xml_text(
+               xml_find_first(x, ".//pub-id[@pub-id-type='pmid']"),
+                                                      trim=TRUE), character(1))
+   doi <-  vapply(z, function(x) xml_text(
+               xml_find_first(x, ".//pub-id[@pub-id-type='doi']"),
+                                                      trim=TRUE), character(1))
+   a1 <-   lapply(z, function(x) xml_text(
+               xml_find_all(x, ".//surname"), trim=TRUE))
+   a2 <-   lapply(z, function(x) xml_text(
+              xml_find_all(x, ".//given-names"), trim=TRUE))
+   # if all references have same number of authors, use SIMPLIFY=FALSE,
+   # see PMC6369050
+   authors <- vapply(mapply(paste, a1,a2, SIMPLIFY=FALSE),
+                          function(x) paste(x, collapse=", "), character(1))
    authors[authors == ""]<-NA
    # same authors published twice in same year, 2012a 2012b
-   # year <-  sapply(z, function(x) xml_integer(xml_find_first(x, ".//year")))
-   year <-  sapply(z, function(x) xml_text(xml_find_first(x, ".//year"), trim=TRUE))
+   year <-  vapply(z, function(x) xml_text(
+               xml_find_first(x, ".//year"), trim=TRUE), character(1))
    if(all(grepl("^[0-9]+$", year))) year <- as.integer(year)
-   title <- sapply(z, function(x) xml_text( xml_find_first(x, ".//article-title"), trim=TRUE))
+   title <- vapply(z, function(x) xml_text(
+             xml_find_first(x, ".//article-title"), trim=TRUE), character(1))
    # new lines in title PMC4909105
    title <- gsub("\n *", " ", title)
-   journal<-sapply(z, function(x) xml_text( xml_find_first(x, ".//source"), trim=TRUE))
-   volume <-sapply(z, function(x) xml_text( xml_find_first(x, ".//volume"), trim=TRUE))
-   p1 <-    sapply(z, function(x) xml_text( xml_find_first(x, ".//fpage"), trim=TRUE))
[…]
