+++ R/pmc_xml.R
@@ -1,7 +1,7 @@
 #' Download PMC XML
 #'
-#' A wrapper for \code{epmc_ftxt} in the \code{europepmc} package to download an
-#' \code{xml_document} from Europe PMC
+#' A wrapper for \code{epmc_ftxt} in the \code{europepmc} package to download
+#' an \code{xml_document} from Europe PMC
 #'
 #' @param id a PMC id starting with 'PMC'
 #'
