### 2019-04-09 f491aa9
vapply errors
