### 2019-04-09 #290 @cstubben
I just updated the package and fixed the `gp()` suggestions except two.  I added tests for all functions but still get warnings?   Also, I would like to leave the dependency to `xml2` since the package parses an `xml_document` and users should know how to view nodes, for example queries like this are highly recommended: `cat(as.character(xml_find_all(doc, "//ref")[1]))`.  More details on xml2 parsing are in the vignette.

Here's the lastest `gp()`

```
It is good practice to

 ✖ write unit tests for all functions, and all package code in general. 77% of code lines are covered by
   test cases.

   R/collapse_rows.R:22:NA
   R/collapse_rows.R:36:NA
   R/collapse_rows.R:40:NA
   R/collapse_rows.R:46:NA
   R/path_string.R:23:NA
   ... and 116 more lines

 ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with other
   packages. Use "Imports" instead.
 ✖ not import packages as a whole, as this can cause name clashes between the imported packages.
   Instead, import only the specific functions you need.
```
