### 2019-04-09 #290 @noamross
Thank you, quite an improvement!  A few notes:

- The "loading packages as a whole message" is a matter of judgement - we just try to avoid importing a whole package when only a few functions are needed.  I will leave it to reviewers to say if there are specific dependencies that can be streamlined.
- re: `xml2`, even if you use it in a vignette and its complementary to your package, you should still leave the package in `Imports` (or `Suggest` if you don't depend on it otherwise) and leave it to the user to load it to use those functions.
- 77% test coverage is reasonable.  We ask the reviewers to inspect the package tests to ensure that appropriate functionality is tested.  However, this is also one reason why it is we require that you have continuous integration with test status and coverage badges set up - so reviewers can inspect the results.  See this chapter in our guide for instructions: https://ropensci.github.io/dev_guide/ci.html .  Once this is done we can proceed.
