+++ R/pmc_table.R
@@ -14,7 +14,8 @@
 #'
 #' @examples
 #' # doc <- europepmc::epmc_ftxt("PMC2231364")
-#' doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
+#' doc <- xml2::read_xml(system.file("extdata/PMC2231364.xml", 
+#'         package = "tidypmc"))
 #' x <- pmc_table(doc)
 #' sapply(x, dim)
 #' x
@@ -26,33 +27,33 @@ pmc_table  <- function(doc){
    if(class(doc)[1] != "xml_document"){
         stop("doc should be an xml_document from PMC")
    }
-   twn  <- length(xml_find_all(doc, "//table-wrap"))
+   twn  <- length(xml2::xml_find_all(doc, "//table-wrap"))
    ## Avoid table-wrap without table node .. link to image only
-   z  <- xml_find_all(doc, "//table-wrap/table/..")
+   z  <- xml2::xml_find_all(doc, "//table-wrap/table/..")
    if (length(z) == 0) {
       message("No tables found")
       if(twn > 0) message("Table-wrap with link to image?")
       tbls <- NULL
    }else{
-        tbl_nodes <- xml_find_all(z, "./table")
+        tbl_nodes <- xml2::xml_find_all(z, "./table")
         message("Parsing ", length(z), " tables")
         if(twn > length(z)){
             message(twn-length(n), " /table-wrap with link to image?")
         }
         ## START table function
-        #  t1 <- xml_find_all(doc, "//table")[1]
+        #  t1 <- xml2::xml_find_all(doc, "//table")[1]
         tbls <- lapply(tbl_nodes, function(t1){
            #PARSE HEADER
-           x <- xml_find_all(t1, ".//thead/tr")
+           x <- xml2::xml_find_all(t1, ".//thead/tr")
            # cat(as.character(x))
            ## missing header
            if(length(x) == 0){
               thead<-NA
            ## 1 header row...
            }else if(length(x) == 1 ){
-              colspan <- as.numeric( xml_attr(
-                     xml_find_all(x, ".//td|.//th"), "colspan", default="1"))
-              thead <- xml_text( xml_find_all(x, ".//td|.//th"))
+              colspan <- as.numeric( xml2::xml_attr(
+                  xml2::xml_find_all(x, ".//td|.//th"), "colspan", default="1"))
+              thead <- xml2::xml_text( xml2::xml_find_all(x, ".//td|.//th"))
               # repeat across colspan
               if( any(colspan>1) ){
                 thead <- rep(thead, colspan)
@@ -61,16 +62,17 @@ pmc_table  <- function(doc){
            # SEE  tables 1 and 2 in PMC3109299
            }else{
                nr <- length(x)
-               nc <- max(vapply(x, function(y) sum(as.numeric( xml_attr(
-                     xml_find_all(y, ".//td|.//th"), "colspan", default="1"))),
-                       double(1)))
+               nc <- max(vapply(x, function(y) sum(as.numeric( xml2::xml_attr(
+                        xml2::xml_find_all(y, ".//td|.//th"), "colspan",
+                         default="1"))), double(1)))
                c2 <- data.frame(matrix(NA, nrow = nr , ncol =  nc ))
                for (i in seq_len(nr)){
-                  rowspan <- as.numeric( xml_attr(
-                   xml_find_all(x[[i]], ".//td|.//th"), "rowspan", default="1"))
-                  colspan <- as.numeric( xml_attr(
-                   xml_find_all(x[[i]], ".//td|.//th"), "colspan", default="1"))
-                  thead <- xml_text( xml_find_all(x[[i]], ".//td|.//th"))
+                  rowspan <- as.numeric( xml2::xml_attr(xml2::xml_find_all(
[…]
