+++ tests/testthat/tests-pmc_table.R
@@ -1,7 +1,7 @@
 context("Parse tables")
 
-doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
-doc2 <- read_xml("<p>This is some text</p>")
+doc <-xml2::read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
+doc2 <- xml2::read_xml("<p>This is some text</p>")
 t1 <- pmc_table(doc)
 
 
