+++ R/pmc_xml.R
@@ -14,8 +14,6 @@
 #' doc <- europepmc::epmc_ftxt("PMC2231364") # OR
 #' doc <- pmc_xml("PMC2231364")
 #' }
-#' doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
-#' doc
 #'
 #' @export
 
