### 2019-04-10 #290 @cstubben
Ok, I fixed the Imports issue and added Travis CI and coverage badge.
