+++ vignettes/pmcftp.md
@@ -1,0 +1,93 @@
+Parsing Europe PMC FTP files
+================
+Chris Stubben
+April 11, 2019
+
+The [Europe PMC FTP](https://europepmc.org/ftp/oa/) includes 2.4 million open access articles separated into files with 10K articles each. Download and unzip a recent series of PMC ids and load into R using the `readr` package.
+
+``` r
+library(readr)
+# all 10K articles or first 100 articles for vignette
+# pmc <- read_lines("PMC6358576_PMC6373241.xml")
+pmcfile <- system.file("extdata/PMC6358576_PMC6358715.xml", package = "tidypmc")
+pmc <- read_lines(pmcfile)
+```
+
+Find the start of the articles nodes.
+
+``` r
+a1 <- grep("^<article ", pmc)
+head(a1)
+#  [1]  2 30 38 52 62 69
+n <- length(a1)
+n
+#  [1] 100
+```
+
+Read a single article by collapsing the lines into a new line separated string.
+
+``` r
+library(xml2)
+x1 <- paste(pmc[2:29], collapse="\n")
+doc <- read_xml(x1)
+doc
+#  {xml_document}
+#  <article article-type="case-report" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:mml="http://www.w3.org/1998/Math/MathML">
+#  [1] <front>\n  <journal-meta>\n    <journal-id journal-id-type="nlm-ta">ACG Case Rep J</journal-i ...
+#  [2] <body>\n  <sec sec-type="intro" id="sec1">\n    <title>Introduction</title>\n    <p>Bezoars a ...
+#  [3] <back>\n  <ref-list>\n    <title>References</title>\n    <ref id="B1">\n      <label>1.</labe ...
+```
+
+Loop through the articles and save the metadata and text below. All 10K articles takes about 10 minutes to run on a Mac laptop and returns 1.7M sentences.
+
+``` r
+a1 <- c(a1, length(pmc))
+met1 <- vector("list", n)
+txt1 <- vector("list", n)
+for(i in seq_len(n)){
+  doc <- read_xml(paste(pmc[a1[i]:(a1[i+1]-1)], collapse="\n"))
+  m1 <- pmc_metadata(doc)
+  id <- m1$PMCID
+  message("Parsing ", i, ". ", id)
+  met1[[i]] <- m1
+  txt1[[i]] <- pmc_text(doc)
+}
+```
+
+Combine the list of metadata and text into tables.
+
+``` r
+met <- bind_rows( met1)
+names(txt1) <- met$PMCID
+txt <- bind_rows( txt1, .id="PMCID")
+met
+#  # A tibble: 100 x 12
+#     PMCID Title Authors  Year Journal Volume Pages `Published onli… `Date received` DOI   Publisher
+#     <chr> <chr> <chr>   <int> <chr>   <chr>  <chr> <chr>            <chr>           <chr> <chr>    
+#   1 PMC6… Endo… Dana B…  2018 ACG Ca… 5      e87   2018-12-5        2018-7-8        10.1… American…
+#   2 PMC6… Chro… Scott …  2018 ACG Ca… 5      e94   2018-12-5        2018-5-5        10.1… American…
+#   3 PMC6… Bile… Steffi…  2018 ACG Ca… 5      e88   2018-12-5        2018-5-7        10.1… American…
+#   4 PMC6… New … Gordon…  2018 ACG Ca… 5      e92   2018-12-5        2018-3-3        10.1… American…
+#   5 PMC6… Bile… Michae…  2018 ACG Ca… 5      e89   2018-12-5        2017-11-3       10.1… American…
[…]
