### 2019-04-11 #290 @cstubben
Thanks, I added that badge and a second vignette.  I don't plan to make any more changes until reviews are returned
