### 2019-04-30 #290 @toph-allen
I've got a start on mine here: [toph-allen/tidypmc-review ](https://github.com/toph-allen/tidypmc-review) (trying out the `pkgreviewr` package for this one). I should have no problem getting it done by Thursday.
