### 2019-05-02 #290 @toph-allen
Here's my review! Additionally, the output of `pkgreviewr`'s index.Rmd notebook, including the results of tests and stuff, is viewable [here](https://toph-allen.github.io/tidypmc-review/index.nb.html). Nice work, @cstubben!

@noamross & @cstubben, please let me know if there are any further questions about what I've written.

---

In summary: Looks good for the most part! Installation instructions, vignettes, and documentation for exported functions are present. I think some minor edits to the README, giving just a little more by way of introduction and context to the package's use cases, and adding contribution guidelines, would help it conform a bit more to the packaging guidelines.

The DESCRIPTION's `Title` — and in general, wherever that sort of introductory piece of text is used without context, e.g. the `R/tidypmc-package.R` help` — should spell out "PubMed Central" rather than using the acronym "PMC".

The GitHub Issues page is given for `BugReports` in the DESCRIPTION, but there are no contributing guidelines in the README or a separate CONTRIBUTING file.

The documentation for all the core functions is comprehensive and contains examples, both in the package's help and in the README file.

The README gives a comprehensive walkthrough of the package's features and examples of their usage, and points the user to the vignettes for further information. I think it could do with a slightly longer introduction, though. Per the first item in the above list, it could describe more of the high-level goals that a user could accomplish with it, or explain a little more how it might fit into users' workflows.

For example, it could describe why a user might want to parse an XML file from the PMC OAS, and why this package is better at that than a generic XML parser.

While the vignettes explain how to work on raw XML files, the README only mentions using the `europepmc` package, and I initially wondered if this was the only way it could read files (I experimented with `xml2` before I discovered the documentation in the vignettes). It might be good to just note, where the README mentions using `europepmc`, that users can also read XML files directly using `xml2`, and refer the user to the vignettes there.

The "Introduction to tidypmc" vignette describes how the package finds and replaces superscript references — which is really clever — but skimming this documentation, it sounded like the package didn't do this automatically. Perhaps move the sentence saying that `pmc_text()` adds brackets to superscript references to the start of that section of the vignette.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

As far as I can tell, the package mostly conforms to the ROpenSci packaging guidelines.

Here are a few random thoughts from my run-through of the package and read-through of the code:

The files main exported functions — those prefixed with `pmc_` and `separate_` are functional and well-documented, and they parsed XML files I threw at them as I would have expected, given the documentation.

The functions `repeat_sub()` and `path_string()` shouldn't be exported, as they are unrelated to the user-facing functionality and serve internal functions.

The package probably doesn't need the function `pmc_xml()`, which is an alias to `europepmc::epmc_ftxt()`, especially since the README file mention the latter function and not the former.

The main functions are covered by tests.

Code is pretty understandable at a readthrough. There are a few areas where one or two comments might help to make clear the purpose of a block, or a loop, where variable names etc. aren't as understandable at a glance.

There's one "TO DO" comment in `collapse_rows.R` (line 43) — not sure if this is something which is still pending or not.

This is pretty nitpicky now, but, while the code style is consistent throughout, there are one or two areas where spacing before and after assignment operators is a little wonky, or there was less spacing around parentheses and brackets in if/else statements — but again, this is *super* nitpicky and somewhat a matter of personal taste. 😅

Though there are tests to cover most functions, `goodpractice` still flags some lines as not being covered. I don't have that much familiarity with coverage metrics — perhaps there are some cases or outcomes which aren't being tested? I'm not sure.

Estimated hours spent reviewing: Roughly 4 hours, including setup, reviewing guidelines, and reviewing the package.

—-

As far as the specific review criteria go, I think this package is really far along in terms of meeting them. A few edits to the documentation, to add a little more context, and resolving the various other minor issues I mentioned, and I think it's good to go.

I think this package also has some clear points of expansion for a next major version. The parsing functions all work very well — but it might be nice to have functions which, for example, extract the body text of an article into a character vector. This sort of thing would do a lot to make the package useful for a larger range of workflows.

NI know I spent most of my writing time here finding nitpicks and critiquing, so I think I should end on a positive note: Nice work! Working with the PMC OAS can be a hassle at times, and this sort of tool
makes it much more accessible.
