### 2019-05-03 #290 @cstubben
@toph-allen Thanks for the great suggestions and careful review.  I'll have time later next week to make these changes.  The introduction definitely needs work too
