### 2019-05-05 #290 @juliasilge
Here is my review! Thank you 🙏 for your patience as I took a few extra days to get this finished.

This package exhibits good design principles and has been built well to create analysis-ready datasets. 🎉 I think the work put into the API design of the functions is especially excellent.

I am an outsider to PubMed and what the target audience for this field of research is, and I think a little bit more context in the README and vignettes would be helpful here. When I visited the GitHub page for [europepmc](https://github.com/ropensci/europepmc), I got a bit more of an understanding of what the problems being solved with this software are.

You might consider switching to [remotes](https://github.com/r-lib/remotes) from devtools for the installation instructions, as remotes is a smaller, more lightweight package without compiled code or dependencies.

For the README and vignettes, I recommend checking out the [tidyverse style guide for pipes](https://style.tidyverse.org/pipes.html) and aligning the way piping is used and formatted with that, or at least so that it is all consistent within this package (pipes go to a new line at the very least, etc). Having a package that explicitly embraces tidy data principles like this one is even better for users when it can align the rest of an ecosystem, in terms of how code is used.

This package is missing a description of contribution guidelines and a code of conduct. This does not need to be long or extensive. You can check [a way that I have written that here](https://github.com/juliasilge/tidytext#community-guidelines). 👥

Are the functions `repeat_sub()` and `path_string()` meant to be exported functions for the user? It seems like probably not? I would either show how these are meant to be used in a vignette, or else don't export them. Here is [what a bunch of documented but not exported functions](https://github.com/ropensci/qualtRics/blob/master/R/utils.R) look like in a package I maintain, although I forgot to add `#' @noRd` there.

One of the XML files used in a vignette is not small, and causes R CMD check to now give a NOTE:

>❯ checking installed package size ... NOTE
>    installed size is  8.3Mb
>    sub-directories of 1Mb or more:
>      extdata   8.0Mb
>
>0 errors ✔ | 0 warnings ✔ | 1 note ✖

It's PMC6358576_PMC6358715.xml that is the big one, at 7.7Mb. Is it possible to use something smaller? Unfortunately you'll need to get it dramatically smaller to get under 1Mb. 😬 It's not generally good practice to have such large data for vignettes, although I do understand about how useful it can be to use real world data to demonstrate with. I would suggest rethinking this in some way because having this NOTE only for a vignette doesn't seem like a battle worth fighting to get this package on CRAN. I think you should make a smaller dataset to use in vignettes but another options that exists is [precompiling a vignette](https://community.rstudio.com/t/precompiling-vignette-with-devtools/1583/5).

Have you tried out the [styler](https://github.com/r-lib/styler) package to clean up and harmonize the code style of this package? It's pretty painless and can improve the overall readability.

In `DESCRIPTION`, your title should be "Title Case" and also should probably spell out what PMC stands for. Your description should **not** start out with "This package..." Your author field [should look like this](https://github.com/ropensci/qualtRics/blob/master/DESCRIPTION), using the `Authors@R` format.

---

It seems like from the title of the package that you are wanting to embrace some tidyverse style and principles. You already have stringr imported in this package, so you might consider using stringr functions like `str_detect()` in place of `grep()` in vignettes, etc. These functions tends to be more readable, more sensible order of arguments, etc.

Also, you may want to consider avoiding loops 🚫 for iteration and instead use purrr, if a goal is to have this package connect to the rest of the suite of tidy tools. For example, in the pmcftp vignette, replace the loop with:

```
tibble(a1) %>%
  mutate(a2 = lead(a1) - 1) %>%
  filter(!is.na(a2)) %>%
  mutate(doc = map2(a1, a2, ~ paste(pmc[.x:.y],
                                    collapse = "\n")),
         docxml = map(doc, read_xml),
         m1 = map(docxml, pmc_metadata),
         txt1 = map(docxml, pmc_text)) %>%
  transmute(PMCID = map_chr(m1, "PMCID"),
            txt1) %>%
  unnest()
```

This code replaces both the chunk called `loop` and the one called `combine` (I have assigned it to anything, but it gives you the idea).

This package is shaping up nicely! ✨ I am really happy to see the excellent work put into this package that will make these datasets so much more easily available to folks who want to access and analyze the content in these texts.
