### 2019-05-06 #290 @cstubben
@juliasilge Thanks for the review and all the helpful links.  I will likely make updates this weekend and use a precompiled vignette to load all 10K articles instead of the 7.7Mb file with 100 articles.
