+++ R/pmc_text.R
@@ -3,152 +3,167 @@
 #' Split section paragraph tags into a table with subsection titles and
 #' sentences using \code{tokenize_sentences}
 #'
-#' @param doc \code{xml_document} from PMC
+#' @param doc \code{xml_document} from PubMed Central
 #'
 #' @return a tibble with section, paragraph and sentence number and text
 #'
 #' @note Subsections may be nested to arbitrary depths and this function will
-#' return the entire path to the subsection title as a delimited string, eg.
-#' "Results; Predicted functions; Pathogenicity".  Any tables, figures and
-#' formulas that are nested in section paragraphs are removed.
+#' return the entire path to the subsection title as a delimited string like
+#' "Results; Predicted functions; Pathogenicity".  Tables, figures and
+#' formulas that are nested in section paragraphs are removed, superscripted
+#' references are replaced with brackets, and any other superscripts or
+#' subscripts are separared with ^ and _.
 #'
 #' @author Chris Stubben
 #'
 #' @examples
-#' # doc <- europepmc::epmc_ftxt("PMC2231364")
+#' # doc <- pmc_xml("PMC2231364")
 #' doc <- xml2::read_xml(system.file("extdata/PMC2231364.xml",
-#'   package = "tidypmc"))
+#'   package = "tidypmc"
+#' ))
 #' txt <- pmc_text(doc)
 #' txt
-#' dplyr::count(txt, section, sort=TRUE)
-#'
+#' dplyr::count(txt, section, sort = TRUE)
 #' @export
 
-pmc_text <- function(doc){
-   if(class(doc)[1] != "xml_document"){
-        stop("doc should be an xml_document from PMC")
-   }
-   ## create new document to remove nodes
-   doc2 <- xml2::xml_new_root(doc)
-   z <- vector("list")
-   ## Main title
-   t1 <- xml2::xml_text(xml2::xml_find_first(doc2,
-                       "//front//article-title"), trim=TRUE)
-   if(!is.na(t1)) z[["Title"]] <- t1
-   ## Abstract
-   a1  <- xml2::xml_text(xml2::xml_find_all(doc2,
-                       "//abstract[not(@abstract-type='summary')]//p"))
-   if(length(a1) > 0) z[["Abstract"]] <- a1
+pmc_text <- function(doc) {
+  if (class(doc)[1] != "xml_document") {
+    stop("doc should be an xml_document from PubMed Central")
+  }
+  ## create new document to remove nodes
+  doc2 <- xml2::xml_new_root(doc)
+  z <- vector("list")
+  ## Main title
+  t1 <- xml2::xml_text(xml2::xml_find_first(
+    doc2, "//front//article-title"
+  ), trim = TRUE)
+  if (!is.na(t1)) z[["Title"]] <- t1
+  ## Abstract
+  a1 <- xml2::xml_text(xml2::xml_find_all(
+    doc2, "//abstract[not(@abstract-type='summary')]//p"
+  ))
+  if (length(a1) > 0) z[["Abstract"]] <- a1
 
-   ## Author summary
-   author_sum <- xml2::xml_text(xml2::xml_find_all(doc2,
-                         "//abstract[@abstract-type='summary']/title"))
-   if(length(author_sum) > 0){
-      z[[author_sum]] <- xml2::xml_text(xml2::xml_find_all(doc2,
-                           "//abstract[@abstract-type='summary']//p"))
-   }
-   if(length(z)== 0){
-       message("No title or abstract found.  Not a PMC XML document?")
-       x <- NULL
-   }else{
-      ## check for tables, figures, formula within <sec/p> tags
-      n <-  xml2::xml_find_all(doc2, "//sec/p/table-wrap")
-      if(length(n) > 0){
[…]
