### 2019-06-02 #290 @cstubben
Sorry for the delays, making plans to work on weekends rarely works out.   I tried to address all the suggestions from both reviews and updated the README introduction to fit the statement of need, added community guidelines, updated code using `styler`,  fixed the exports, tidyverse pipe style, PMC abbreviations, title case and the other issues above except two -  I left devtools in the README (never used or installed `remotes`, but maybe I should) and still need to precompile the second FTP vignette and remove the 7.7 Mb file next week.
