### 2019-06-02 652b8a2
add rmarkdown for Travis CI
