+++ DESCRIPTION
@@ -22,6 +22,7 @@ Imports:
     readr
 Suggests:
     tidytext,
+    rmarkdown,
     knitr,
     testthat,
     covr
