### 2019-06-02 #290 @noamross
Thank you for the update, @cstubben. @juliasilge and @toph-allen, please let us know if the changes address your comments.

@cstubben I do note that `devtools::install_github()` just calls `remotes::install_github()`, so you have used and installed it. 😉
