### 2019-06-10 #290 @juliasilge
This is looking great! I especially think the updated README is improved a lot. 🎉

From my perspective, things that still definitely need to happen are:

- In the `DESCRIPTION` file, the `Description` field cannot start with "This package..."
- The vignette needs to be updated so it does not use data too large for this package to go to CRAN.

I also still recommend using **remotes** instead of **devtools** for installation instructions, because of how heavy devtools is.
