### 2019-06-10 #290 @cstubben
Thanks! I fixed the Description and added **remotes** to the README.   I already reduced the large XML file from the FTP site to 1/10th the previous size, so files in inst/extdata are less than 1 MB combined.
