### 2019-06-10 #290 @juliasilge
So @cstubben do you no longer intend to precompile a vignette?
