### 2019-06-10 #290 @cstubben
No.  I mainly wanted to show how to find the start of the article nodes and read the combined XML files on the FTP site.  I might use a pre-compiled vignette later to read all 2.5 million articles and look at a few use cases like mentions of "statistically significant" across different journals, but will likely need some collaborators 😄
