### 2019-06-14 #290 @juliasilge
From my perspective, this package is ready to be approved! I'm glad to see this sample of the life science published literature so much more accessible. 🙌
