### 2019-06-24 #290 @toph-allen
Sorry it took me a while to finish reviewing the updates.

Looking through the changes, I agree that the package is ready to go! The updated README does a much better job of introducing the topics of the packages, as well as clearly documenting all the package's core functionality. The improvements to the code, both in terms of formatting consistency and minor implementation changes, are also good.

I don't have any additional comments at this time. Nice work! 👏
