### 2019-06-24 #290 @noamross
Thank you @toph-allen and @juliasilge!  @cstubben, one tiny thing I found in my final checks: It appears that you have the **europepmc** in `Imports` but it is only used in the vignette, so it should be in `Suggests:`.   With that little fix, we are ready to approve!

To-dos:
- [ ] Update any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`.
- [ ] Update your `codemeta.json`

Our developer guide has a [section](https://ropensci.github.io/dev_guide/collaboration.html) for ongoing maintenance after acceptance. Welcome aboard!
