+++ README.Rmd
@@ -9,8 +9,8 @@ knitr::opts_chunk$set(
 )
 ```
 
-[![Build Status](https://travis-ci.org/cstubben/tidypmc.svg?branch=master)](https://travis-ci.org/cstubben/tidypmc)
-[![Coverage status](https://codecov.io/gh/cstubben/tidypmc/branch/master/graph/badge.svg)](https://codecov.io/github/cstubben/tidypmc?branch=master)
+[![Build Status](https://travis-ci.org/ropensci/tidypmc.svg?branch=master)](https://travis-ci.org/ropensci/tidypmc)
+[![Coverage status](https://codecov.io/gh/ropensci/tidypmc/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/tidypmc?branch=master)
 [![](https://badges.ropensci.org/290_status.svg)](https://github.com/ropensci/software-review/issues/290)
 
 # tidypmc
@@ -31,7 +31,7 @@ sections, paragraphs, sentences or table rows.
 Use [remotes] to install the package.
 
 ```{r install, eval=FALSE}
-remotes::install_github("cstubben/tidypmc")
+remotes::install_github("ropensci/tidypmc")
 ```
 
 ## Load XML
@@ -180,7 +180,7 @@ has details on parsing XML files at the Europe PMC [FTP site].
 This project is released with a [Contributor Code of Conduct](CONDUCT.md). By
 participating in this project you agree to abide by its terms. Feedback, bug
 reports, and feature requests are welcome
-[here](https://github.com/cstubben/tidypmc/issues).
+[here](https://github.com/ropensci/tidypmc/issues).
 
 
 [remotes]: https://github.com/r-lib/remotes
@@ -190,8 +190,8 @@ reports, and feature requests are welcome
 [FTP site]: https://europepmc.org/ftp/oa/
 [tidytext]: https://www.tidytextmining.com/
 [stringr]: https://stringr.tidyverse.org/
-[vignette]: https://github.com/cstubben/tidypmc/blob/master/vignettes/tidypmc.md
-[PMC FTP vignette]: https://github.com/cstubben/tidypmc/blob/master/vignettes/pmcftp.md
+[vignette]: https://github.com/ropensci/tidypmc/blob/master/vignettes/tidypmc.md
+[PMC FTP vignette]: https://github.com/ropensci/tidypmc/blob/master/vignettes/pmcftp.md
 [tokenizers]: https://lincolnmullen.com/software/tokenizers/
 [xml2]: https://github.com/r-lib/xml2
 [europepmc]: https://github.com/ropensci/europepmc
