### 2019-06-24 #290 @cstubben
Thanks for all your help!  I fixed the Imports, transferred the repo and updated all GitHub links from cstubben to ropensci.
